+++ R/MODIS-package.R
@@ -20,7 +20,9 @@
 #' @importFrom graphics abline box grid locator title
 #' @importFrom mapedit selectFeatures
 #' @importFrom maps map.axes
-#' @importFrom methods as
+#' @importFrom maptools map2SpatialPolygons map2SpatialLines checkPolygonsHoles
+#' @importFrom methods as slot
+#' @importFrom rgeos gBuffer gIsValid
 #' @importFrom sf st_as_sf
 #' @importFrom stats na.omit smooth.spline
 #' @importFrom utils capture.output download.file installed.packages read.table vi write.csv write.table